checkstyle fixes
